add the feature of expiration by btree

+++ server.go
@@ -8,6 +8,7 @@ import (
 )
 
 
+
 type Item struct {
 	// Key is the Item's key (250 bytes maximum).
 	Key string
@@ -16,11 +17,12 @@ type Item struct {
 	// Expiration is the cache expiration time, in seconds: either a relative
 	// time from now (up to 1 month), or an absolute Unix epoch time.
 	// Zero means the Item has no expiration time.
-	Expiration int32
+	Expiration int64
 	casid uint64
 }
 
 func main() {
+
 	log.SetFlags(log.Ldate|log.Lshortfile)
 	h:=flag.String("h","127.0.0.1","server ip")
 	p:=flag.String("p","11211","server port")
